+++ capa/render/verbose.py
@@ -33,6 +33,7 @@ import capa.features.freeze as frz
 import capa.render.result_document as rd
 from capa.rules import RuleSet
 from capa.engine import MatchResults
+from capa.capabilities.extract_domain_and_ip import verbose_extract_domain_and_ip
 
 
 def format_address(address: frz.Address) -> str:
@@ -316,6 +317,42 @@ def render_rules(ostream, doc: rd.ResultDocument):
         ostream.writeln(rutils.bold("no capabilities found"))
 
 
+def render_domain_and_ip(ostream: rutils.StringIO, doc: rd.ResultDocument):
+    """
+    example::
+        +-----------------------------------------------------------+
+        | IP addresses and web domains                              |
+        |-----------------------------------------------------------+
+        | google.com                                                |
+        |    |----IP address:                                       |
+        |            |----192.0.0.1                                 |
+        |    |----Functions used to communicate with google.com:    |
+        |            |----InternetConnectA                          |
+        |            |----HttpOpenRequestA                          |
+        |            |----FtpGetFileA                               |
+        |    |----3 occurrances                                     |
+        |                                                           |                                                                          |
+        | 192.123.232.08                                            |
+        |    |----Functions used to communicate with 192.123.232.08:|
+        |            |----...                                       |
+        |                                                           |
+        +-----------------------------------------------------------+
+    """
+    rows = []
+    for domain_or_ip in verbose_extract_domain_and_ip(doc):
+        rows.append(domain_or_ip)
+
+    if rows:
+        ostream.write(
+            tabulate.tabulate(
+                {"IP addresses and web domains": rows}, headers=["keys"], tablefmt="mixed_outline",
+            )
+        )
+        ostream.write("\n")
+    else:
+        ostream.writeln(rutils.bold("No web domains or IP addresses found"))
+
+
 def render_verbose(doc: rd.ResultDocument):
     ostream = rutils.StringIO()
 
@@ -323,6 +360,9 @@ def render_verbose(doc: rd.ResultDocument):
     ostream.write("\n")
 
     render_rules(ostream, doc)
+    ostream.write("\n")
+
+    render_domain_and_ip(ostream, doc)
     ostream.write("\n")
 
     return ostream.getvalue()