+++ capa/render/vverbose.py
@@ -22,6 +22,7 @@ import capa.render.result_document as rd
 import capa.features.freeze.features as frzf
 from capa.rules import RuleSet
 from capa.engine import MatchResults
+from capa.capabilities.extract_domain_and_ip import verbose_extract_domain_and_ip
 
 logger = logging.getLogger(__name__)
 
@@ -458,6 +459,42 @@ def render_rules(ostream, doc: rd.ResultDocument):
         ostream.writeln(rutils.bold("no capabilities found"))
 
 
+def render_domain_and_ip(ostream: rutils.StringIO, doc: rd.ResultDocument):
+    """
+    example::
+        +------------------------------------------------------------+
+        | IP addresses and web domains                               |
+        |------------------------------------------------------------+
+        | google.com                                                 |
+        |    |----IP address:                                        |
+        |            |----192.0.0.1                                  |
+        |    |----Functions used to communicate with google.com:     |
+        |            |----InternetConnectA                           |
+        |            |----HttpOpenRequestA                           |
+        |            |----FtpGetFileA                                |
+        |    |----3 occurrances                                      |
+        |                                                            |                                                                          |
+        | 192.123.232.08                                             |
+        |    |----Functions used to communicate with 192.123.232.08: |
+        |            |----...                                        |
+        |                                                            |
+        +------------------------------------------------------------+
+    """
+    rows = []
+    for domain_or_ip in verbose_extract_domain_and_ip(doc):
+        rows.append(domain_or_ip)
+
+    if rows:
+        ostream.write(
+            tabulate.tabulate(
+                {"IP addresses and web domains": rows}, headers=["keys"], tablefmt="mixed_outline",
+            )
+        )
+        ostream.write("\n")
+    else:
+        ostream.writeln(rutils.bold("No web domains or IP addresses found"))
+
+
 def render_vverbose(doc: rd.ResultDocument):
     ostream = rutils.StringIO()
 
@@ -465,6 +502,9 @@ def render_vverbose(doc: rd.ResultDocument):
     ostream.write("\n")
 
     render_rules(ostream, doc)
+    ostream.write("\n")
+
+    render_domain_and_ip(ostream, doc)
     ostream.write("\n")
 
     return ostream.getvalue()